% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casuality.R
\name{causality_classification}
\alias{causality_classification}
\title{Causality classification}
\usage{
causality_classification(hypothesis_df)
}
\arguments{
\item{hypothesis_df}{hypothesis statement output of \code{\link[=hypothesis_extraction]{hypothesis_extraction()}}}
}
\description{
Wrapper function. Executes all steps in the causality classification process.
The causality classification model was trained under the following
conditions :
\itemize{
\item Token normalization method: stemming
\item Cause/Effect entity replacement: yes
\item Imbalanced Sampling: no
\item Feature processing: Bag-of-words
\item Model: Naive Bayes - Multinomial
}
}
