% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{compile_table}
\alias{compile_table}
\title{Compile final table}
\usage{
compile_table(hypothesis, entities, causality, direction, file_name)
}
\arguments{
\item{hypothesis}{Output of \code{\link[=hypothesis_extraction]{hypothesis_extraction()}}}

\item{entities}{Output of \code{\link[=entity_extraction]{entity_extraction()}}}

\item{causality}{Output of \code{\link[=causality_classification]{causality_classification()}}}

\item{direction}{Output of \code{\link[=direction_classification]{direction_classification()}}}

\item{file_name}{File name of processed PDFs}
}
\description{
Compiles all calculated values into single table
}
