% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{direction_classification}
\alias{direction_classification}
\title{Direction classification}
\usage{
direction_classification(hypothesis_df)
}
\arguments{
\item{hypothesis_df}{hypothesis statement output of \code{\link[=hypothesis_extraction]{hypothesis_extraction()}}}
}
\description{
Wrapper function. Executes all steps in the direction classification process.
The causality classification model was trained under the following
conditions :
\itemize{
\item Token normalization method: stemming
\item Cause/Effect entity replacement: no
\item Imbalanced Sampling: no
\item Feature processing: Bag-of-words
\item Model: Support Vector Machines
}
}
