% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{drop_empty_hypothesis}
\alias{drop_empty_hypothesis}
\title{Drop empty hypothesis sentences}
\usage{
drop_empty_hypothesis(hypothesis_index, unique_hypothesis_label, input_text)
}
\arguments{
\item{hypothesis_index}{Vector listing indexes of each hypothesis}

\item{unique_hypothesis_label}{Vector of unique hypothesis labels}

\item{input_text}{Processed input text, one sentence per vector element}
}
\description{
Removes sentences in input text which only contain the hypothesis tag. No
other text is contained in that sentence.
}
