% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{drop_hypothesis_below_min_threshold}
\alias{drop_hypothesis_below_min_threshold}
\title{Drop hypothesis sentences with fewer than minimum token threshold}
\usage{
drop_hypothesis_below_min_threshold(input_text, min_threshold = 6)
}
\arguments{
\item{input_text}{Processed input text, one sentence per vector element}

\item{min_threshold}{Minimum threshold of tokens in a sentence.}
}
\description{
Removes sentences that contain a hypothesis tag, and contain fewer than
the minimum threshold of tokens. This is a method to assist in removing
erroneous hypothesis identification.
}
