% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{hypothesis_extraction}
\alias{hypothesis_extraction}
\title{Extract Hypotheses
The following function accepts processed text in character vector form
and returns hypothesis statements.
Extract hypothesis statements}
\usage{
hypothesis_extraction(input_text, apply_model = TRUE)
}
\arguments{
\item{input_text}{PDF text as processed by \code{\link[=process_text]{process_text()}}.}

\item{apply_model}{Boolean tag for whether to filter hypothesis statements
with the hypothesis classification model.}
}
\description{
Wrapper function. Executes all steps in the hypothesis extraction process.
}
