% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{process_text}
\alias{process_text}
\title{Process PDF text}
\usage{
process_text(input_path)
}
\arguments{
\item{input_path}{path to PDF file}
}
\description{
Wrapper function. Executes all steps in the process flow converting raw
PDF file into text. This processing step returns text that will be used as
input for the three major functions in the package:
\itemize{
\item Hypothesis classification
\item Entity extraction
\item Causality classification
}
}
