% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{unique_hypothesis_labels}
\alias{unique_hypothesis_labels}
\title{Reduce to unique hypothesis labels}
\usage{
unique_hypothesis_labels(hypothesis_labels)
}
\arguments{
\item{hypothesis_labels}{Vector of identified hypothesis labels}
}
\description{
Reduces the list of identified hypotheses to unique labels. This also
includes dropping any numeric hypothesis label where an alphanumeric label
with the same number has appeared earlier in the document, i.e.: If
Hypothesis 1 appears after Hypothesis 1a, Hypothesis 1 is removed
}
