% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity.R
\name{entity_extraction}
\alias{entity_extraction}
\title{Extract entity clauses (multiple cases)}
\usage{
entity_extraction(hypothesis_df)
}
\arguments{
\item{hypothesis_df}{hypothesis statement output of \code{\link[=hypothesis_extraction]{hypothesis_extraction()}}}
}
\description{
Wrapper function. Executes all steps in the entity  extraction process for
a multiple hypothesis statements.
}
